Moved conversion methods to extension methods
