+++ TRENZ.Lib.RazorMail.MailKit/MailKitMailClient.cs
@@ -1,5 +1,4 @@
-﻿using System.Linq;
-using System.Threading;
+﻿using System.Threading;
 using System.Threading.Tasks;
 
 using JetBrains.Annotations;
@@ -17,8 +16,6 @@ using TRENZ.Lib.RazorMail.Models;
 using TRENZ.Lib.RazorMail.Services;
 
 using RazorMailMessage = TRENZ.Lib.RazorMail.Models.MailMessage;
-using MailKitMailMessage = MimeKit.MimeMessage;
-using IMailKitMailMessage = MimeKit.IMimeMessage;
 
 namespace TRENZ.Lib.RazorMail;
 
@@ -40,56 +37,17 @@ public class MailKitMailClient(IOptions<SmtpAccount> accountOptions, ILogger<Mai
     /// <inheritdoc />
     protected override async Task SendInternalAsync(RazorMailMessage message, CancellationToken cancellationToken)
     {
-        var nativeMessage = ConvertToNative(message);
+        var mimeMessage = message.ToMimeMessage();
 
         using var client = await CreateClientAsync(cancellationToken);
 
         var formatOptions = FormatOptions.Default.Clone();
 
         logger.LogInformation("Sending mail from {From} to {Recipients} (CC: {Cc}, BCC: {Bcc}) with subject {Subject}",
-            nativeMessage.From, nativeMessage.To, nativeMessage.Cc, nativeMessage.Bcc, nativeMessage.Subject);
+            mimeMessage.From, mimeMessage.To, mimeMessage.Cc, mimeMessage.Bcc, mimeMessage.Subject);
 
-        await client.SendAsync(formatOptions, nativeMessage, cancellationToken);
+        await client.SendAsync(formatOptions, mimeMessage, cancellationToken);
 
         await client.DisconnectAsync(true, cancellationToken);
-    }
-
-    private static MailKitMailMessage ConvertToNative(RazorMailMessage razorMessage)
-    {
-        var mailKitMessage = new MailKitMailMessage();
-
-        SetMailHeaders(razorMessage, mailKitMessage);
-        SetMailContent(razorMessage, mailKitMessage);
-
-        return mailKitMessage;
-    }
-
-    private static void SetMailHeaders(RazorMailMessage razorMessage, IMailKitMailMessage mailMessage)
-    {
-        mailMessage.From.Add(razorMessage.Headers.From!.ToMailboxAddress());
-        mailMessage.To.AddRange(razorMessage.Headers.Recipients.Select(x => x.ToMailboxAddress()));
-        mailMessage.Cc.AddRange(razorMessage.Headers.CarbonCopy.Select(x => x.ToMailboxAddress()));
-        mailMessage.Bcc.AddRange(razorMessage.Headers.BlindCarbonCopy.Select(x => x.ToMailboxAddress()));
-        mailMessage.ReplyTo.AddRange(razorMessage.Headers.ReplyTo.Select(x => x.ToMailboxAddress()));
-    }
-
-    private static void SetMailContent(RazorMailMessage razorMessage, IMailKitMailMessage mailMessage)
-    {
-        mailMessage.Subject = razorMessage.Content.Subject;
-
-        var bodyBuilder = new BodyBuilder
-        {
-            HtmlBody = razorMessage.Content.HtmlBody,
-        };
-
-        foreach (var item in razorMessage.Content.Attachments.Values)
-        {
-            var attachment =
-                bodyBuilder.Attachments.Add(item.FileName, item.FileData, ContentType.Parse(item.ContentType));
-            attachment.ContentId = item.ContentId;
-            attachment.IsAttachment = !item.Inline;
-        }
-
-        mailMessage.Body = bodyBuilder.ToMessageBody();
     }
 }
